django_s3_app/requirements.txt
# Django and Core Dependencies
Django==5.0.9
djangorestframework==3.15.2
django-cors-headers==4.4.0

# AWS MSK (Kafka) Integration
kafka-python==2.0.2

# AWS SDK and S3
boto3==1.35.36
botocore==1.35.36

# Authentication and JWT
PyJWT==2.9.0
requests==2.32.3

# Environment and Configuration
python-decouple==3.8

# Development and Testing (optional)
pytest==8.3.3
pytest-django==4.9.0
coverage==7.6.1

# Production WSGI Server (optional)
gunicorn==23.0.0
whitenoise==6.8.1

# Logging and Monitoring (optional)
sentry-sdk==2.15.0

# Utilities
python-dateutil==2.9.0
pytz==2024.2